third_party/blink/web_tests/platform/generic/http/tests/inspector-protocol/fenced-frame/setWebLifecycleState-fenced-frame.https-expected.txt

Tests that Page.setWebLifecycleState() in a fenced frame is not allowed.
Page.setWebLifecycleState() from a fenced frame:
PASS: This is only supported for top-level frames
